RF: some minor reorganizations of imports within main __init__.py -- hiding things away from public view
